+++ radiospectra/spectrogram/sources/rpw.py
@@ -16,11 +16,11 @@ class RPWSpectrogram(GenericSpectrogram):
     >>> from radiospectra.spectrogram import Spectrogram
     >>> query = Fido.search(a.Time('2020-07-11', '2020-07-11 23:59'), a.Instrument('RPW'),
     ...             a.Level(2), a.soar.Product('RPW-HFR-SURV')) #doctest: +REMOTE_DATA
-    >>> downloaded = Fido.fetch(query[0])  #doctest: +REMOTE_DATA
-    >>> spec = Spectrogram(downloaded[0])  #doctest: +REMOTE_DATA
-    >>> spec  #doctest: +REMOTE_DATA
+    >>> downloaded = Fido.fetch(query[0])  #doctest: +SKIP
+    >>> spec = Spectrogram(downloaded[0])  #doctest: +SKIP
+    >>> spec  #doctest: +SKIP
     [<RPWSpectrogram SOLO, RPW, RPW-AGC1 375.0 kHz - 16375.0 kHz, 2020-07-11T00:00:39.352 to 2020-07-12T00:00:55.715>, <RPWSpectrogram SOLO, RPW, RPW-AGC2 375.0 kHz - 16375.0 kHz, 2020-07-11T00:00:39.352 to 2020-07-12T00:00:55.715>]
-    >>> spec[0] .plot()  #doctest: +REMOTE_DATA
+    >>> spec[0] .plot()  #doctest: +SKIP
     <matplotlib.collections.QuadMesh object at ...>
     """