+++ CHANGELOG.rst
@@ -47,4 +47,4 @@ Features
 --------
 
 - Add Parker Solar Probe (PSP) Radio Frequency Receiver (RFS) Fido client `radiospectra.net.sources.psp.RFSClient`. (`#34 <https://github.com/sunpy/radiospectra/pull/34>`__)
-- Add STEREO WAVES (SWAVES) Fido client `radiospectra.net.SWAVESClient`. (`#35 <https://github.com/sunpy/radiospectra/pull/35>`__)
+- Add STEREO WAVES (SWAVES) Fido client ``radiospectra.net.SWAVESClient``. (`#35 <https://github.com/sunpy/radiospectra/pull/35>`__)